+++ bundle/src/main/java/com/adobe/acs/commons/mcp/form/MultifieldComponent.java
@@ -19,10 +19,6 @@
  */
 package com.adobe.acs.commons.mcp.form;
 
-import com.adobe.acs.commons.mcp.util.AnnotatedFieldDeserializer;
-import java.lang.reflect.ParameterizedType;
-import java.util.LinkedHashMap;
-import java.util.Map;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceMetadata;
 
@@ -31,44 +27,26 @@ import org.apache.sling.api.resource.ResourceMetadata;
  * where this is used, some javascript should be included by the front-end to
  * process the resulting form correctly.
  */
-public class MultifieldComponent extends FieldComponent {
-
-    Map<String, FieldComponent> fieldComponents = new LinkedHashMap<>();
-    boolean isComposite = true;
+public class MultifieldComponent extends AbstractContainerComponent {
 
     public MultifieldComponent() {
         setResourceType("granite/ui/components/coral/foundation/form/multifield");
     }
 
-    public void init() {
-        if (getField() != null) {
-            ParameterizedType type = (ParameterizedType) getField().getGenericType();
-            Class clazz = (Class) type.getActualTypeArguments()[0];
-            extractFieldComponents(clazz);
-        }
-        if (sling != null && sling.getRequest() != null) {
-            setPath(sling.getRequest().getResource().getPath());
-        }
-    }
-
-    @Override
+     @Override
     public Resource buildComponentResource() {
-        getComponentMetadata().put("composite", isComposite);
+        getComponentMetadata().put("composite", isComposite());
         AbstractResourceImpl res = new AbstractResourceImpl(getPath(), getResourceType(), getResourceSuperType(), getComponentMetadata());
         if (sling != null) {
             res.setResourceResolver(sling.getRequest().getResourceResolver());
         }
-        if (isComposite) {
+        if (isComposite()) {
             AbstractResourceImpl field = new AbstractResourceImpl(getPath() + "/field", "granite/ui/components/coral/foundation/container", getResourceSuperType(), new ResourceMetadata());
             // The container component is what sets the name, not the base component
             field.getResourceMetadata().put("name", getName());
             res.addChild(field);
-            AbstractResourceImpl items = new AbstractResourceImpl(getPath() + "/field/items", "", "", new ResourceMetadata());
+            AbstractResourceImpl items = generateItemsResource(getPath() + "/field", true);
             field.addChild(items);
-            for (FieldComponent component : fieldComponents.values()) {
-                component.setPath(getPath() + "/field/items/" + component.getName());
-                items.addChild(component.buildComponentResource());
-            }
         } else {
             for (FieldComponent component : fieldComponents.values()) {
                 component.setPath(getPath() + "/field");
@@ -79,26 +57,5 @@ public class MultifieldComponent extends FieldComponent {
         }
 
         return res;
-    }
-
-    public Map<String, FieldComponent> getFieldComponents() {
-        return fieldComponents;
-    }
-
-    private void extractFieldComponents(Class clazz) {
-        fieldComponents = new LinkedHashMap<>();
-        if (clazz == String.class) {
-            FieldComponent comp = new TextfieldComponent();
-            FormField fieldDef = FormField.Factory.create(getName(), "", null, false, comp.getClass(), null);
-            comp.setup(getName(), null, fieldDef, sling);
-            comp.getComponentMetadata().put("title", getName());
-            // TODO: Provide a proper mechanism for setting path when creating components
-            fieldComponents.put(getName(), comp);
-            isComposite = false;
-        } else {
-            fieldComponents.putAll(AnnotatedFieldDeserializer.getFormFields(clazz, sling));
-            isComposite = true;
-        }
-        fieldComponents.values().forEach(this::addClientLibraries);
     }
 }