+++ bundle/src/test/java/com/adobe/acs/commons/mcp/form/FieldComponentTest.java
@@ -104,6 +104,11 @@ public class FieldComponentTest {
                 }
 
                 @Override
+                public String category() {
+                    return null;
+                }
+
+                @Override
                 public boolean required() {
                     return false;
                 }